requirements.txt
Django==1.7.4
MySQL-python==1.2.5
django-filter==0.10.0
djangorestframework==3.1.3
flup==1.0.2
wsgiref==0.1.2
